ensure ArchConfiguration is reset after data point generation
Adjusting `ArchConfiguration` in a static context without resetting it afterwards is a bad idea. Together with `ArchConfigurationRule` it then depends on the order which default value for `resolveMissingDependenciesFromClasspath` is stored within `ArchConfigurationRule`. This is a recipe for flaky tests, e.g. on master `ClassFileImporterTest` succeeds, while on this branch it fails, even though there have been no relevant changes there.
To make it easier to execute little snippets and correctly ensure the configuration is reset before the snippet as well as after the snippet, I have introduced a new utility method `ArchConfigurationRule.resetConfigurationAround(callable)`.

Signed-off-by: Peter Gafert <peter.gafert@tngtech.com>