use instance scoped `BoundProcessors` instead of a single static one
This was a little premature optimization. I assumed creating a new `BoundProcessor` for every `JavaClass` would be an unnecessary overhead. On the other hand having a single shared instance has downsides for maintainability and a higher danger to introduce bugs. When I tried to measure performance I could not detect any difference between having a single static `BoundProcessor` and instantiating one for every class, so we should go for the instance scoped one.
Also the instance scoped pattern is easier to apply to the superclass case, so it makes sense to refactor this first.

Signed-off-by: Peter Gafert <peter.gafert@tngtech.com>