force use of list in `TypeParametersBuilder`
Since the order of type parameters matters we should not allow a generic `Collection` to be passed to the `TypeParametersBuilder` and then convert it to a list, but instead force a list to be passed right away. The only other type ever used was an empty set, so this has not caused any problems so far, but for future maintenance I have fixed this now.

Signed-off-by: Peter Gafert <peter.gafert@tngtech.com>