fix signature of `Optional.or(Optional<..> value)`
Since `Optional` is covariant, it is a perfectly valid use case to have `Optional.of(parent).or(optionalChild)`, since an `Optional.of(child)` is always also an `Optional.of(parent)` (for an inheritance relation `Child extends Parent`). Unfortunately the signature so far did not reflect that, because `Optional.of(parent)` only allowed another `Optional<Parent>` to be passed to `Optional.or(..)`. This commit fixes this to allow any optional of a subtype of the type parameter to be passed into `Optional.or(..)`.

Signed-off-by: Peter Gafert <peter.gafert@tngtech.com>