add generic superclass to `JavaClass`
Unfortunately signature handling is a little bit inconsistent. If the type is not generic the signature is `null` and we need to rely on the superclass name passed to `visit(..)`. If the type is generic it seems we could derive the full superclass information from the generic signature of the class. However I'm not completely sure, if this is the case for all scenarios. So I decided to play it safe and implement this backwards compatible, i.e. the raw superclass information is derived exactly as it has, if we also have a class signature containing a generic superclass, then the information within `JavaClass` will be enriched with a `JavaParameterizedType genericSuperclass` that can be retrieved through the new method `JavaClass.getGenericSuperclass()`. This is consistent with the Reflection API as far as I can see, same as that `getGenericSuperclass()` will return a "raw" `JavaClass`, if the type is not generic.

Signed-off-by: Peter Gafert <peter.gafert@tngtech.com>